+++ packages/scaffold-config/src/frameworks/index.ts
@@ -1,0 +1,21 @@
+import { crav4 } from './crav4'
+import { crav5 } from './crav5'
+import { nextjs } from './nextjs'
+import { nuxtjs } from './nuxtjs'
+import { react } from './react'
+import { vue2 } from './vue2'
+import { vue3 } from './vue3'
+import { vueclivue2 } from './vueclivue2'
+import { vueclivue3 } from './vueclivue3'
+
+export const FRONTEND_FRAMEWORKS = [
+  crav4,
+  crav5,
+  nuxtjs,
+  nextjs,
+  react,
+  vue2,
+  vue3,
+  vueclivue3,
+  vueclivue2,
+] as const
