+++ packages/scaffold-config/src/frameworks/isWebpack.ts
@@ -1,0 +1,3 @@
+import type { Bundler } from '../types'
+
+export const isWebpack = (bundler: Bundler) => ['webpack4', 'webpack5'].includes(bundler)
